Address PR review: split modules, top-level imports, XML safety
- Split discovery logic into sorter_sample_discovery.py (path discovery,
  XML parsing, sample loading) and keep report HTML in
  sorter_multi_sample_report.py
- Move boto3 and cloud_sync imports to top level (fail gracefully)
- Wrap XML parsing in try-except (warn, don't crash)
- Add as_dataframe kwarg to read_sorter_statistics_csv
- Replace "ppmSeq" filter with "applicationQC"
- Early-stop iteration when all XML samples found
- Remove dead code branch in _assemble_multi_html
